Add x64 build targets for all XRVessels.sln projects
Also:
* Cleaned up x64 compile warnings
* Added build type to version string (32-bit or 64-bit)
* Updated email address in file comment headers (headers will be updated again shortly anyway for open-source license)

+++ XRVessels/DeltaGliderXR1/XR1Lib/XR1ConfigFileParser.cpp
@@ -632,7 +632,7 @@ void XR1ConfigFileParser::ApplyCheatcodesIfEnabled()
     {
         if (GetCheatcodesFoundCount() > 0)
         {
-            sprintf(msg, "*** WARNING: %d CHEATCODE(S) set but ignored: cheatcodes are disabled (check 'CheatcodesEnabled' setting)", m_cheatcodeVector.size()); 
+            sprintf(msg, "*** WARNING: %d CHEATCODE(S) set but ignored: cheatcodes are disabled (check 'CheatcodesEnabled' setting)", static_cast<int>(m_cheatcodeVector.size()));
             WriteLog(msg);
         }
     }
